+++ src/test/kotlin/choliver/nespot/sixfiveohtwo/BitwiseTest.kt
@@ -1,4 +1,4 @@
-package choliver.nespot.sixfiveohtwo.instructions
+package choliver.nespot.sixfiveohtwo
 
 import choliver.nespot.sixfiveohtwo.assertForAddressModes
 import choliver.nespot.sixfiveohtwo.model.Opcode.*
